Move DummyHandler into testlib to allow removal of the dynamic discovery mechanism from unit tests entirely
