+++ aodncore/testlib/__init__.py
@@ -1,9 +1,11 @@
 from .basetest import BaseTestCase
+from .dummyhandler import DummyHandler
 from .handlertest import HandlerTestCase
 from .testutil import (dest_path_testing, get_nonexistent_path, get_test_config, make_test_file, make_zip, mock)
 
 __all__ = [
     'BaseTestCase',
+    'DummyHandler',
     'HandlerTestCase',
     'dest_path_testing',
     'get_nonexistent_path',
